feat: config 병합
